Break out the prerender code into smaller parts
