+++ packages/prerender/src/browser.js
@@ -1,0 +1,5 @@
+// This file is here, so that browser utils can be imported
+// like this improt {isBrowser} from '@redwoodjs/prerender/browser'
+// Workaround to to TS not suppoting package.json exports
+
+export * from './browser.utils'
